24 DONE - somente usuários autorizados podem deletar threads

+++ app/Http/Controllers/ThreadsController.php
@@ -106,7 +106,13 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
+        
+        if (request()->wantsJson()) {
+            return response([], 204);
+        }
 
         return redirect('/threads');
     }
